hms-enterprise-grade/services/consent/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
prometheus-fastapi-instrumentator==7.0.0
SQLAlchemy==2.0.36
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.10
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
